Refactor Lock to be counting (for rate limiting)
